requirements.txt
pyaudio
wave
azure-cognitiveservices-speech
python-dotenv
openai
pydash
tenacity
azure-ai-textanalytics
